requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
PyOpenGL==3.1.10
PyQt5==5.15.11
trimesh==4.8.2
Pillow>=10.0.0
pyglet<2
networkx>=3.0
scipy>=1.10.0




